requirements.txt
certifi==2018.1.18
chardet==3.0.4
clearbit==0.1.7
Django==2.0.2
django-filter==1.1.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
idna==2.6
Markdown==2.6.11
Pillow==5.0.0
PyJWT==1.5.3
pytz==2018.3
requests==2.18.4
urllib3==1.22
